apps/api/requirements.txt
# LangGraph + LangChain
langchain-openai>=0.3.0
langgraph>=0.4.0
langgraph-checkpoint-postgres>=0.1.0
langchain-core>=0.3.0

# Web framework
fastapi[standard]
uvicorn[standard]

# Database
psycopg[binary]>=3.0
asyncpg>=0.30.0

# HTTP client (for JSearch API)
httpx

# Web search
tavily-python

# File I/O
aiofiles

# Config
pyyaml

# Scheduling
apscheduler>=3.10.0,<4.0

# Settings
pydantic-settings
python-dotenv

# DNA system (embeddings + similarity)
openai>=1.0
numpy>=1.24
